requirements.txt
# Phase 1 - Data Collection & Preprocessing
# PDF text extraction
pdfplumber>=0.10.0
pdfminer.six>=20221105

# OCR
pytesseract>=0.3.10
pdf2image>=1.16.3
Pillow>=10.0.0

# Text processing
nltk>=3.8.1
scikit-learn>=1.3.0

# Data handling
pandas>=2.0.0
numpy>=1.24.0

# Serialization
joblib>=1.3.0
scipy>=1.11.0

# Utilities
python-dateutil>=2.8.2